+++ gftools/kmesh.hpp
@@ -8,7 +8,7 @@
 #include "grid_base.hpp"
 #include "almost_equal.hpp"
 
-namespace gftools { 
+namespace gftools {
 
 ///this class implements a regular equidistant grid, typically used as a regular k-space grid between 0 and 2PI.
 ///lower boundary is always zero
@@ -25,9 +25,9 @@ public:
     kmesh():npoints_(0){};
     ///constructor from a vector of regularly spaced ints
     kmesh(std::vector<real_type> const& in);
-   
+
     ///given a double 'in', it finds that double in the grid.
-    ///The first return value is whether it has been found or not 
+    ///The first return value is whether it has been found or not
     ///The second return value is the index.
     ///The third return value 'weight' measures how close the point is to a grid point
     std::tuple <bool, size_t, real_type> find(real_type in) const ;
@@ -35,8 +35,8 @@ public:
     template <class Obj> auto integrate(const Obj &in) const -> typename std::result_of<Obj(decltype(vals_[0]))>::type;
     template <class Obj> auto eval(Obj &in, real_type x) const ->decltype(in[0]);
     template <class Obj> auto eval(Obj &in, point x) const ->decltype(in[0]) { return base::eval(in,x); }
-   
-    ///shift implements a periodic operator+ in three variants 
+
+    ///shift implements a periodic operator+ in three variants
     real_type shift(real_type in,real_type shift_arg) const;
     point shift(point in, real_type shift_arg) const { return static_cast<const base*>(this)->shift(in,shift_arg); }
     point shift(point in, point shift_arg) const { return static_cast<const base*>(this)->shift(in,shift_arg); }
@@ -64,8 +64,8 @@ inline kmesh::kmesh(std::vector<real_type> const& in):
     if (vals_.size() <= 1) throw gftools::ex_generic("Can't construct a kmesh with <= 1 element");
     double diff = vals_[1] - vals_[0];
     bool ok = true;
-    for (int i=1; i<vals_.size() && ok; i++) { 
-         ok = almost_equal(vals_[i] - vals_[i-1], diff, diff * 1e-6); 
+    for (size_t i=1; i<vals_.size() && ok; i++) {
+         ok = almost_equal(vals_[i] - vals_[i-1], diff, diff * 1e-6);
         }
     if (!ok) throw gftools::ex_generic("kmesh should be uniform");
     domain_len_ = vals_.size() * diff;
@@ -78,7 +78,7 @@ inline std::tuple <bool, size_t, real_type> kmesh::find(real_type in) const
     assert(in>=0 && in < domain_len_);
     int n = std::lround(in/domain_len_*npoints_);
     if (n<0) { ERROR("kmesh point is out of bounds, " << in << "<" << 0); return std::make_tuple(false,0,0); };
-    if (n==npoints_) n=0; 
+    if (n==npoints_) n=0;
     if (n>npoints_) { ERROR("kmesh point is out of bounds, " << in << ">" << domain_len_); return std::make_tuple(false,npoints_,0); };
     real_type weight=in/domain_len_*npoints_-real_type(n);
     return std::make_tuple (true,n,weight);
@@ -86,17 +86,17 @@ inline std::tuple <bool, size_t, real_type> kmesh::find(real_type in) const
 
 
 template <class Obj>
-inline auto kmesh::eval(Obj &in, real_type x) const ->decltype(in[0]) 
+inline auto kmesh::eval(Obj &in, real_type x) const ->decltype(in[0])
 {
 	auto p = find_nearest(x);
     return eval(in,p);
 }
 
-template <class Obj> 
+template <class Obj>
 auto kmesh::integrate(const Obj &in) const -> typename std::result_of<Obj(decltype(vals_[0]))>::type
 {
     decltype(in(vals_[0])) R = in(real_type(vals_[0]));
-    R=std::accumulate(vals_.begin()+1, vals_.end(), R,[&](decltype(in(vals_[0]))& y,decltype(vals_[0]) & x) {return y+in(x);}); 
+    R=std::accumulate(vals_.begin()+1, vals_.end(), R,[&](decltype(in(vals_[0]))& y,decltype(vals_[0]) & x) {return y+in(x);});
     return R/npoints_;
 }
 
@@ -104,7 +104,7 @@ inline real_type kmesh::shift(real_type in, real_type shift_arg) const
 {
     assert (in>=0 && in < domain_len_);
     real_type out;
-    out = in + real_type(shift_arg); 
+    out = in + real_type(shift_arg);
     out-= domain_len_*(almost_equal(out, domain_len_, num_io<double>::tolerance())?1.0:std::floor(out/domain_len_));
     return out;
 }
